add objects; align `assign`
Couple comments:
For now in this language there is no ambiguity about what is meant when,
and I especially noticed that `name()` calls and object creation could be
unambiguous using either `=` or `:` (or both).

Conversely the syntax for `name`d functions can be understood as
- `name(a:3,b:4)` meaning "we are creating an object `{a:3,b:4}`
  and using that as the context for the function call";
- `name(a=3,b=4)` meaning "we are creating a context `a=3,b=4` then
  evaluating the (postponed) expression stored in `name` in that context
  (i.e. as in `a=3, b=4, <expression>`).
The second syntax looks closer to the semantics; the first syntax closer
to what is used in other languages.